requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-environ==0.11.2
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
whitenoise==6.7.0
